# specs/001-ai-rag-eval-platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI RAG-Based Document Evaluation & QnA Platform API
  description: API for document upload, RAG-based QnA, and automated evaluation
  version: 0.1.0
  contact:
    name: ELP Gemini Team
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /admin/dashboard:
    get:
      tags: [admin]
      summary: Admin dashboard metrics
      description: Get aggregate platform metrics
      operationId: getAdminDashboard
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDashboardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users:
    get:
      tags: [admin]
      summary: List all users
      description: Get list of platform users
      operationId: listAllUsers
      security:
        - sessionAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUserResponse'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{user_id}:
    get:
      tags: [admin]
      summary: Get user details
      description: Retrieve specific user information and activity
      operationId: getUserDetails
      security:
        - sessionAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/qna-logs:
    get:
      tags: [admin]
      summary: List QnA logs
      description: Get all QnA interactions across platform
      operationId: listQALogs
      security:
        - sessionAuth: []
      parameters:
        - name: user_id
          in: query
          schema:
            type: integer
        - name: document_id
          in: query
          schema:
            type: integer
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: QnA logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  qa_logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminQALogResponse'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/prompts:
    get:
      tags: [admin]
      summary: List system prompts
      description: Get all system prompts with versions
      operationId: listSystemPrompts
      security:
        - sessionAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [qna, evaluation]
      responses:
        '200':
          description: System prompts
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompts:
                    type: array
                    items:
                      $ref: '#/components/schemas/SystemPromptResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [admin]
      summary: Create system prompt
      description: Create new version of system prompt
      operationId: createSystemPrompt
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, content]
              properties:
                type:
                  type: string
                  enum: [qna, evaluation]
                content:
                  type: string
                  maxLength: 10000
                is_active:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Prompt created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemPromptResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/prompts/{prompt_id}/activate:
    post:
      tags: [admin]
      summary: Activate system prompt
      description: Set prompt version as active
      operationId: activateSystemPrompt
      security:
        - sessionAuth: []
      parameters:
        - name: prompt_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Prompt activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Prompt activated successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
...
  schemas:
...
    DocumentResponse:
      type: object
      properties:
        id:
          type: integer
          example: 42
        title:
          type: string
          example: My Research Paper
        status:
          type: string
          enum: [uploading, indexing, ready, failed, deleted]
          example: ready
        file_size:
          type: integer
          example: 2457600
        uploaded_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    QALogResponse:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        answer:
          type: string
        model_name:
          type: string
        created_at:
          type: string
          format: date-time

    # ==================== Evaluation Schemas ====================
    EvaluationRunResponse:
      type: object
      properties:
        id:
          type: integer
          example: 5
        status:
          type: string
          enum: [pending, running, success, failed]
          example: success
        template_id:
          type: integer
          example: 1
        model_name:
          type: string
          example: gemini-2.0-flash-thinking-exp
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
...
    AdminDashboardResponse:
      type: object
      properties:
        total_users:
          type: integer
          example: 150
        total_documents:
          type: integer
          example: 1250
        documents_last_7_days:
          type: integer
          example: 75
        total_qna_requests:
          type: integer
          example: 8500
        qna_requests_last_7_days:
          type: integer
          example: 320
        total_evaluation_runs:
          type: integer
          example: 450
        evaluation_runs_last_7_days:
          type: integer
          example: 28
    AdminUserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        is_admin:
          type: boolean
        document_count:
          type: integer
        qna_count:
          type: integer
        created_at:
          type: string
          format: date-time
    AdminUserDetailResponse:
      allOf:
        - $ref: '#/components/schemas/AdminUserResponse'
        - type: object
          properties:
            documents:
              type: array
              items:
                $ref: '#/components/schemas/DocumentResponse'
            recent_qna:
              type: array
              items:
                $ref: '#/components/schemas/QALogResponse'
            recent_evaluations:
              type: array
              items:
                $ref: '#/components/schemas/EvaluationRunResponse'
    AdminQALogResponse:
      allOf:
        - $ref: '#/components/schemas/QALogResponse'
        - type: object
          properties:
            user_id:
              type: integer
            document_id:
              type: integer
            document_title:
              type: string
            prompt_version:
              type: integer
    SystemPromptResponse:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [qna, evaluation]
        version:
          type: integer
        content:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ==================== Error Schemas ====================
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: Error description
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...